aider: Update image path in `service-card-r.board.tsx` to use a path that can be used in an HTML `img` tag.
# Aider chat conversation:

USER: convert this to path that html img tag can take "D:\Dev\codux\Blaide\src\assets\img1.webp"

ASSISTANT: To convert a local file system path to a path that can be used in an HTML `img` tag, you need to serve the image from a location that is accessible via HTTP. Typically, this means placing the image within the public directory of your web server or application.

Assuming that the `D:\Dev\codux\Blaide\src\assets\` directory is accessible via the root of your server and that your server is configured to serve files from the `assets` directory, the path you would use in an HTML `img` tag would be `/assets/img1.webp`.

However, if you need to update the path in the `service-card-r.board.tsx` file to reflect this change, here is the diff:

```diff
--- src/_codux/boards/service-card-r/service-card-r.board.tsx
+++ src/_codux/boards/service-card-r/service-card-r.board.tsx
@@ ... @@
 export default createBoard({
     name: 'ServiceCardR',
     Board: () => <ServiceCardR
                     className="example-class"
                     heading="Sample Heading"
                     description="This is a sample description for the ServiceCardR component."
-                    img="D:/Dev/codux/Blaide/src/assets/img1.webp"
+                    img="/assets/img1.webp"
                  />,
     isSnippet: true,
 });
```

This diff assumes that the `assets` directory is at the root of your server's public directory. If the actual path differs, you will need to adjust the `img` property accordingly.