aider: Update the `img` URL string to a relative path in `service-card-r.board.tsx`.
# Aider chat conversation:

USER: in `src\_codux\boards\service-card-r\service-card-r.board.tsx` iam unable to view the image mention within img prop. the absolute path of image is D:\Dev\codux\Blaide\src\assets\img1.webp. how should I update the img url string ?

ASSISTANT: To update the `img` URL string in `service-card-r.board.tsx` to a relative path that can be used in a web application, you should change the absolute path to a relative path that is relative to the location of the `service-card-r.board.tsx` file within your project structure. Assuming that the `assets` folder is in the `src` directory and is served statically by your web server, the relative path from `service-card-r.board.tsx` to `img1.webp` would be `../../../assets/img1.webp`.

Here is the diff to update the `img` property:

```diff
--- src/_codux/boards/service-card-r/service-card-r.board.tsx
+++ src/_codux/boards/service-card-r/service-card-r.board.tsx
@@ ... @@
 export default createBoard({
     name: 'ServiceCardR',
     Board: () => <ServiceCardR
                     className="example-class"
                     heading="Sample Heading"
                     description="This is a sample description for the ServiceCardR component."
-                    img="D:/Dev/codux/Blaide/src/assets/img1.webp"
+                    img="../../../assets/img1.webp"
                  />,
     isSnippet: true,
 });
```

This change will make the image path relative, which should allow it to be resolved correctly when the web application is running.