aider: Set the background color of the `HomePage` component to white.
# Aider chat conversation:

USER: why is the background on HomePage all changed, it was earlier white i think

ASSISTANT: To investigate the issue with the background color on the `HomePage` component, I would need to look at the relevant CSS file, which is `home-page.module.scss`. If you would like me to proceed with checking and potentially fixing the background color issue, please add the `home-page.module.scss` file to the chat so I can make the necessary changes.

USER: I added these *read-write* files: src\components\home-page\home-page.module.scss

ASSISTANT: To ensure the `HomePage` component has a white background as it was earlier, I will provide a diff to update the `home-page.module.scss` file. This will set the background color of the root element to white.

```diff
--- src\components\home-page\home-page.module.scss
+++ src\components\home-page\home-page.module.scss
@@ ... @@
 .root {
     font: $large-font;
     display: flex;
     flex-direction: column;
     row-gap: 0.5em;
-    background-color: $snow-white;
+    background-color: #FFFFFF; // Set the background color to white
     padding-top: 2em;
     padding-bottom: 2em;
 }
```

This change will apply a white background color to the `.root` class, which is used by the `HomePage` component. If `$snow-white` was intended to be white and is not being applied correctly, this direct hex color code will ensure the background is white. If `$snow-white` is defined in `theme.module.scss` and is not white, you may want to check that file for the correct value.